requirements.txt
# ══ API Framework ══
fastapi>=0.115
uvicorn[standard]>=0.30
gunicorn>=22
pydantic[email]>=2.7
python-multipart>=0.0.9
grpcio>=1.64
protobuf>=5.0

# ══ Vision & ML ══
torch>=2.3
torchvision>=0.18
ultralytics>=8.2          # YOLOv11
transformers>=4.41         # ViT, CLIP, SAM
opencv-python-headless>=4.10
Pillow>=10.3
imageio>=2.34
scikit-image>=0.23
onnxruntime>=1.18
albumentations>=1.4
timm>=1.0                  # vision backbone zoo

# ══ Async / HTTP ══
httpx[http2]>=0.27
aiofiles>=23.2
anyio>=4.4
ARQ>=0.25                  # async job queue

# ══ Task Queue ══
celery[redis]>=5.4
flower>=2.0                # Celery monitor UI
redis>=5.0

# ══ Data & Storage ══
polars>=0.20               # pandas killer
deltalake>=0.17            # Delta Lake ACID
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29              # PostgreSQL async driver
alembic>=1.13
pyarrow>=16.0

# ══ Performance ══
cython>=3.0
numba>=0.60
numpy>=1.26
scipy>=1.13

# ══ Observability ══
loguru>=0.7
opentelemetry-sdk>=1.24
opentelemetry-instrumentation-fastapi>=0.45b0
prometheus-client>=0.20
tenacity>=8.3              # retry logic

# ══ Dev & Testing ══
pytest>=8.2
pytest-asyncio>=0.23
hypothesis>=6.100          # property-based testing
ruff>=0.4                  # linter ultra rápido (reemplaza flake8+black+isort)
mypy>=1.10
pre-commit>=3.7